+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Versioning follows a `MAJOR`.`MINOR`.`PATCH` pattern, where:
 ### Changed
 
 - Use links and buttons consistently [#258](https://github.com/spraakbanken/mink-frontend/issues/258)
+- Remove table row striping
 
 ## [2.1.1](https://github.com/spraakbanken/mink-frontend/releases/tag/v2.1.1) (2026-06-12)